Revert "Consolidate Empty(If|Else)Block into EmptyConditionBlock"
