+++ Rubberduck.Inspections/Concrete/EmptyIfBlockInspection.cs
@@ -1,0 +1,58 @@
+﻿using Antlr4.Runtime;
+using Antlr4.Runtime.Misc;
+using Rubberduck.Inspections.Abstract;
+using Rubberduck.Inspections.Results;
+using Rubberduck.Parsing;
+using Rubberduck.Parsing.Grammar;
+using Rubberduck.Parsing.Inspections.Abstract;
+using Rubberduck.Parsing.Inspections.Resources;
+using Rubberduck.Parsing.VBA;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Rubberduck.Inspections.Concrete
+{
+    internal class EmptyIfBlockInspection : ParseTreeInspectionBase
+    {
+        public EmptyIfBlockInspection(RubberduckParserState state)
+            : base(state, CodeInspectionSeverity.Suggestion) { }
+
+        public override Type Type => typeof(EmptyIfBlockInspection);
+
+        public override CodeInspectionType InspectionType => CodeInspectionType.CodeQualityIssues;
+
+        public override IEnumerable<IInspectionResult> GetInspectionResults()
+        {
+            return Listener.Contexts
+                .Where(result => !IsIgnoringInspectionResultFor(result.ModuleName, result.Context.Start.Line))
+                .Select(result => new QualifiedContextInspectionResult(this,
+                                                        InspectionsUI.EmptyIfBlockInspectionResultFormat,
+                                                        result));
+        }
+
+        public override IInspectionListener Listener { get; } =
+            new EmptyIfBlockListener();
+
+        public class EmptyIfBlockListener : EmptyBlockInspectionListenerBase
+        {
+            public override void EnterIfStmt([NotNull] VBAParser.IfStmtContext context)
+            {
+                InspectBlockForExecutableStatements(context.block(), context);
+            }
+
+            public override void EnterElseIfBlock([NotNull] VBAParser.ElseIfBlockContext context)
+            {
+                InspectBlockForExecutableStatements(context.block(), context);
+            }
+
+            public override void EnterSingleLineIfStmt([NotNull] VBAParser.SingleLineIfStmtContext context)
+            {
+                if (context.ifWithEmptyThen() != null)
+                {
+                    AddResult(new QualifiedContext<ParserRuleContext>(CurrentModuleName, context.ifWithEmptyThen()));
+                }
+            }
+        }
+    }
+}
