+++ Rubberduck.Inspections/QuickFixes/RemoveEmptyIfBlockQuickFix.cs
@@ -1,23 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using Rubberduck.Inspections.Concrete;
-using Rubberduck.Parsing.VBA;
-using Rubberduck.Parsing.Inspections.Abstract;
-using Rubberduck.Parsing.Grammar;
-using Rubberduck.Parsing.Rewriter;
-using Antlr4.Runtime;
-using Rubberduck.Parsing.Inspections.Resources;
 using System.Diagnostics;
+using System.Linq;
+using Antlr4.Runtime;
+using Rubberduck.Inspections.Concrete;
+using Rubberduck.Parsing.Grammar;
+using Rubberduck.Parsing.Inspections.Abstract;
+using Rubberduck.Parsing.Inspections.Resources;
+using Rubberduck.Parsing.Rewriter;
+using Rubberduck.Parsing.VBA;
 
 namespace Rubberduck.Inspections.QuickFixes
 {
-    internal sealed class RemoveEmptyConditionBlockQuickFix : IQuickFix
+    internal sealed class RemoveEmptyIfBlockQuickFix : IQuickFix
     {
-        private static readonly HashSet<Type> _supportedInspections = new HashSet<Type> { typeof(EmptyConditionBlockInspection) };
+        private static readonly HashSet<Type> _supportedInspections = new HashSet<Type> { typeof(EmptyIfBlockInspection) };
         private readonly RubberduckParserState _state;
 
-        public RemoveEmptyConditionBlockQuickFix(RubberduckParserState state)
+        public RemoveEmptyIfBlockQuickFix(RubberduckParserState state)
         {
             _state = state;
         }
@@ -90,16 +90,6 @@ namespace Rubberduck.Inspections.QuickFixes
             }
 
             rewriter.Remove(context);
-        }
-
-        private void UpdateContext(VBAParser.ElseBlockContext context, IModuleRewriter rewriter)
-        {
-            var elseBlock = context.block();
-
-            if (elseBlock.ChildCount == 0)
-            {
-                rewriter.Remove(context);
-            }
         }
 
         private void UpdateCondition(VBAParser.RelationalOpContext condition, IModuleRewriter rewriter)
@@ -188,7 +178,7 @@ namespace Rubberduck.Inspections.QuickFixes
 
         public string Description(IInspectionResult result)
         {
-            return InspectionsUI.RemoveEmptyConditionBlockQuickFix;
+            return InspectionsUI.RemoveEmptyIfBlockQuickFix;
         }
 
         public bool CanFixInProcedure { get; } = false;
